pyproject.toml
[project]
name = "emg-classification"
version = "0.1.0"
description = "A reusable library for EMG signal classification"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "joblib>=1.3.0",
]

[project.optional-dependencies]
boosting = [
    "xgboost>=2.0.0",
    "lightgbm>=4.0.0",
]
jupyter = [
    "jupyter>=1.0.0",
]
gpu = [
    "thundersvm>=0.3.12",
    "torch>=2.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]
all = [
    "emg-classification[boosting,jupyter,gpu,dev]",
]

[project.scripts]
emg-extract = "emg_classification.cli.extract:main"
emg-train = "emg_classification.cli.train:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/emg_classification"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
